fix: compile sdk without xcode error app extensions (#185)
