+++ Sources/MessagingPush/MessagingPush.swift
@@ -6,35 +6,6 @@ import UIKit
 import UserNotifications
 #endif
 
-public protocol MessagingPushInstance {
-    func registerDeviceToken(_ deviceToken: String)
-    func deleteDeviceToken()
-    func trackMetric(
-        deliveryID: String,
-        event: Metric,
-        deviceToken: String
-    )
-
-    #if canImport(UserNotifications)
-    @discardableResult
-    // sourcery:Name=didReceiveNotificationRequest
-    // sourcery:IfCanImport=UserNotifications
-    func didReceive(
-        _ request: UNNotificationRequest,
-        withContentHandler contentHandler: @escaping (UNNotificationContent) -> Void
-    ) -> Bool
-
-    // sourcery:IfCanImport=UserNotifications
-    func serviceExtensionTimeWillExpire()
-
-    func userNotificationCenter(
-        _ center: UNUserNotificationCenter,
-        didReceive response: UNNotificationResponse,
-        withCompletionHandler completionHandler: @escaping () -> Void
-    ) -> Bool
-    #endif
-}
-
 /**
  Swift code goes into this module that are common to *all* of the Messaging Push modules (APN, FCM, etc).
  So, performing an HTTP request to the API with a device token goes here.
@@ -95,46 +66,4 @@ public class MessagingPush: MessagingPushInstance {
     ) {
         implementation?.trackMetric(deliveryID: deliveryID, event: event, deviceToken: deviceToken)
     }
-
-    #if canImport(UserNotifications)
-    /**
-     - returns:
-     Bool indicating if this push notification is one handled by Customer.io SDK or not.
-     If function returns `false`, `contentHandler` will *not* be called by the SDK.
-     */
-    @discardableResult
-    public func didReceive(
-        _ request: UNNotificationRequest,
-        withContentHandler contentHandler: @escaping (UNNotificationContent) -> Void
-    ) -> Bool {
-        guard let implementation = implementation else {
-            contentHandler(request.content)
-            return false
-        }
-
-        return implementation.didReceive(request, withContentHandler: contentHandler)
-    }
-
-    /**
-     iOS OS telling the notification service to hurry up and stop modifying the push notifications.
-     Stop all network requests and modifying and show the push for what it looks like now.
-     */
-    public func serviceExtensionTimeWillExpire() {
-        implementation?.serviceExtensionTimeWillExpire()
-    }
-
-    public func userNotificationCenter(
-        _ center: UNUserNotificationCenter,
-        didReceive response: UNNotificationResponse,
-        withCompletionHandler completionHandler: @escaping () -> Void
-    ) -> Bool {
-        guard let implementation = implementation else {
-            completionHandler()
-            return false
-        }
-
-        return implementation.userNotificationCenter(center, didReceive: response,
-                                                     withCompletionHandler: completionHandler)
-    }
-    #endif
 }
