+++ Tests/MessagingPushAPN/APITest.swift
@@ -62,10 +62,7 @@ class MessagingPushAPNAPITest: UnitTest {
 
         _ = MessagingPush.shared.userNotificationCenter(.current(), didReceive: UNNotificationResponse.testInstance,
                                                         withCompletionHandler: {})
-        _ = mock.userNotificationCenter(UNUserNotificationCenter.current(),
-                                        didReceive: UNNotificationResponse.testInstance, withCompletionHandler: {})
-        _ = instance.userNotificationCenter(.current(), didReceive: UNNotificationResponse.testInstance,
-                                            withCompletionHandler: {})
+        // Cannot guarantee instance or mock will have userNotificationCenter() function as that function is not available to app extensions.
         #endif
     }
 }
