+++ Tests/Tracking/APITest.swift
@@ -25,7 +25,6 @@ class TrackingAPITest: UnitTest {
 
     // This function checks that public functions exist for the SDK and they are callable.
     // Maybe we forgot to add a function? Maybe we forgot to make a function `public`?
-    // swiftlint:disable:next function_body_length
     func test_allPublicTrackingFunctions() throws {
         try skipRunningTest()
 
@@ -35,8 +34,6 @@ class TrackingAPITest: UnitTest {
 
         // config
         CustomerIO.config { _ in }
-        mock.config { _ in }
-        instance.config { _ in }
 
         // Identify
         CustomerIO.shared.identify(identifier: "")
